--- lib/banka.dart
@@ -1,147 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:fbase/kullanicigiris.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_credit_card/flutter_credit_card.dart';
-
-class CreditCardPage extends StatefulWidget {
-  final email;
-
-  CreditCardPage({Key? mykey, this.email}) : super(key: mykey);
-  @override
-  _CreditCardPageState createState() => _CreditCardPageState();
-}
-
-class _CreditCardPageState extends State<CreditCardPage> {
-
-  String cardNumber = '';
-  String expiryDate = '';
-  String cardHolderName = '';
-  String cvvCode = '';
-  bool isCvvFocused = false;
-  final GlobalKey<FormState> formKey = GlobalKey<FormState>();
-  @override
-  Widget build(BuildContext context) {
-    
-    return Scaffold(
-      body: SafeArea(
-        child: Column(
-          children: [
-            Flexible(
-              child: CreditCardWidget(
-                cardNumber: cardNumber,
-                expiryDate: expiryDate,
-                cardHolderName: cardHolderName,
-                cvvCode: cvvCode,
-                showBackView: isCvvFocused,
-                obscureCardNumber: true,
-                obscureCardCvv: true,
-                onCreditCardWidgetChange: (CreditCardBrand) {},
-              ),
-            ),
-            Expanded(
-                child: SingleChildScrollView(
-              child: Column(
-                children: [
-                  CreditCardForm(
-                    cardNumber: cardNumber,
-                    expiryDate: expiryDate,
-                    cardHolderName: cardHolderName,
-                    cvvCode: cvvCode,
-                    onCreditCardModelChange: onCreditCardModelChange,
-                    themeColor: Colors.red,
-                    formKey: formKey,
-                    cardNumberDecoration: InputDecoration(
-                        border: OutlineInputBorder(),
-                        labelText: 'Number',
-                        hintText: 'xxxx xxxx xxxx xxxx'),
-                    expiryDateDecoration: InputDecoration(
-                        border: OutlineInputBorder(),
-                        labelText: 'Expired Date',
-                        hintText: 'xx/xx'),
-                    cvvCodeDecoration: InputDecoration(
-                        border: OutlineInputBorder(),
-                        labelText: 'CVV',
-                        hintText: 'xxx'),
-                    cardHolderDecoration: InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: 'Card Holder',
-                    ),
-                  ),
-                ],
-              ),
-            )),
-            ElevatedButton(
-              style: ElevatedButton.styleFrom(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8.0),
-                ),
-                backgroundColor: Colors.red,
-              ),
-              child: Container(
-                margin: EdgeInsets.all(8.0),
-                child: Text(
-                  'Save',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontFamily: 'halter',
-                    fontSize: 14,
-                    package: 'flutter_credit_card',
-                  ),
-                ),
-              ),
-              onPressed: () {
-                if (formKey.currentState!.validate()) {
-                  FirebaseFirestore.instance
-                      .collection('Kartlar')
-                      .doc(girismail)
-                      .set({
-                    'isim': cardHolderName,
-                    'kartno': cardNumber,
-                    'limit': 400,
-                    'email': girismail,
-                  });
-                  final snackBar = SnackBar(
-                    content: Container(
-                      width: 150,
-                      height: 50,
-                      child: Center(
-                        child: Text(
-                          'Your card is successfully saved',
-                          style: TextStyle(
-                            fontSize: 18,
-                            
-                          ),
-                        ),
-                        
-                      ),
-                      
-                    ),
-                     action: SnackBarAction(
-                      label: 'Log in again',
-                      onPressed: () {
-                        Navigator.of(context).push(MaterialPageRoute(
-                          builder: (context) => KullaniciGiris(),
-                        ));
-                      },
-                    ),
-                  );
-                  ScaffoldMessenger.of(context).showSnackBar(snackBar);
-                }
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void onCreditCardModelChange(CreditCardModel creditCardModel) {
-    setState(() {
-      cardNumber = creditCardModel.cardNumber;
-      expiryDate = creditCardModel.expiryDate;
-      cardHolderName = creditCardModel.cardHolderName;
-      cvvCode = creditCardModel.cvvCode;
-      isCvvFocused = creditCardModel.isCvvFocused;
-    });
-  }
-}
