+++ lib/changepw.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
-import 'kullanicigiris.dart';
+import 'logging_in/user_logging_in.dart';
 
 class change extends StatefulWidget {
   final email;
