+++ lib/logging_in/google_sign_in.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
-import '../kullaniciekrani.dart';
+import '../user_screen.dart';
 
 class googleSignIn extends StatelessWidget {
   @override
